Aleph 3.2.10

+++ aleph/logic/collections.py
@@ -21,8 +21,9 @@ def create_collection(data, role=None, sync=False):
                                    created_at=created_at)
     if collection.created_at == created_at:
         publish(Events.CREATE_COLLECTION,
-            params={'collection': collection},
-            actor_id=role.id)
+                params={'collection': collection},
+                channels=[collection, role],
+                actor_id=role.id)
     db.session.commit()
     Authz.flush()
     refresh_collection(collection.id)
@@ -57,14 +58,16 @@ def index_collections():
         index.index_collection(collection)
 
 
-def delete_collection(collection, sync=False):
+def delete_collection(collection, keep_metadata=False, sync=False):
     reset_collection(collection, sync=False)
     flush_notifications(collection)
     deleted_at = collection.deleted_at or datetime.utcnow()
     Entity.delete_by_collection(collection.id, deleted_at=deleted_at)
     Document.delete_by_collection(collection.id)
-    Permission.delete_by_collection(collection.id, deleted_at=deleted_at)
-    collection.delete(deleted_at=deleted_at)
+    if not keep_metadata:
+        Permission.delete_by_collection(collection.id, deleted_at=deleted_at)
+        collection.delete(deleted_at=deleted_at)
     db.session.commit()
-    index.delete_collection(collection.id, sync=sync)
-    Authz.flush()
+    if not keep_metadata:
+        index.delete_collection(collection.id, sync=sync)
+        Authz.flush()
